+++ priv/repo/migrations/20241009065244_add_preferences_to_users.exs
@@ -1,0 +1,9 @@
+defmodule Lightning.Repo.Migrations.AddPreferencesToUsers do
+  use Ecto.Migration
+
+  def change do
+    alter table("users") do
+      add :preferences, :jsonb, default: fragment("jsonb_object('{}')"), null: false
+    end
+  end
+end